KG/BATCH-KG-BE/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "batch-kg-fe-backend"
version = "0.1.0"
description = "FastAPI + LangGraph backend for the Spring Batch KG Agent UI"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "mcp>=1.6.0",
    "httpx>=0.27.0",
    "openai>=1.30.0",
    "boto3>=1.34.0",
    "google-generativeai>=0.5.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "config*"]
